requirements.txt
numpy==1.22.3
pydicom==2.3.0
tensorflow==2.10.0
keras-unet-collection==0.1.13
dvc[s3]==2.38.1
dvclive==1.3.0
pandas==1.4.2
scikit-learn==1.2.0
scipy==1.8.0
Pillow==8.3.0
SimpleITK==2.1.1.2
boto3==1.24.59
mlflow==2.2.2
PyYAML==6.0
